refactor:  HSMListV2, and enhance UI components
