+++ src/components/UI/Pager/Pager.module.css
@@ -150,18 +150,18 @@
   flex: 0 0 44px;
   align-items: center;
   justify-content: center;
-  padding: 0;
+  padding: 0 0 0 16px;
 }
 
 .ChevronBtn {
   display: flex;
   align-items: center;
   justify-content: center;
-  width: 28px;
-  height: 28px;
+  width: 36px;
+  height: 36px;
   padding: 0;
   border: none;
-  border-radius: 6px;
+  border-radius: 8px;
   background: none;
   color: #9ca3af;
   cursor: pointer;
@@ -174,8 +174,8 @@
 }
 
 .ChevronBtn svg {
-  width: 14px;
-  height: 14px;
+  width: 22px;
+  height: 22px;
   transition: transform 0.2s ease;
 }
