+++ src/i18n/en/en.json
@@ -793,13 +793,15 @@
   "HSM Template": "HSM Template",
   "Filter by status": "Filter by status",
   "Filter by category": "Filter by category",
-  "Green:": "Green:",
+  "Green": "Green",
   "Approved — the template in this language is approved.": "Approved — the template in this language is approved.",
-  "Yellow:": "Yellow:",
+  "Yellow": "Yellow",
   "Pending — the template in this language is pending.": "Pending — the template in this language is pending.",
-  "Red:": "Red:",
+  "Red": "Red",
   "Rejected / Failed — the template in this language was rejected or failed.": "Rejected / Failed — the template in this language was rejected or failed.",
   "Please wait while we process all the templates": "Please wait while we process all the templates",
+  "Processing your file": "Processing your file",
+  "Bulk apply cancelled. The upload may still finish in the background.": "Bulk apply cancelled. The upload may still finish in the background.",
   "Templates": "Templates",
   "Please go through all the templates below": "Please go through all the templates below",
   "Sync HSM": "Sync HSM",
