+++ src/i18n/hi/hi.json
@@ -652,13 +652,15 @@
   "Select tag": "टैग चुनें",
   "Filter by status": "स्थिति के अनुसार फ़िल्टर करें",
   "Filter by category": "श्रेणी के अनुसार फ़िल्टर करें",
-  "Green:": "हरा:",
+  "Green": "हरा",
   "Approved — the template in this language is approved.": "स्वीकृत — इस भाषा में टेम्पलेट स्वीकृत है।",
-  "Yellow:": "पीला:",
+  "Yellow": "पीला",
   "Pending — the template in this language is pending.": "लंबित — इस भाषा में टेम्पलेट लंबित है।",
-  "Red:": "लाल:",
+  "Red": "लाल",
   "Rejected / Failed — the template in this language was rejected or failed.": "अस्वीकृत / विफल — इस भाषा में टेम्पलेट अस्वीकृत या विफल हो गया।",
   "Please wait while we process all the templates": "कृपया प्रतीक्षा करें जब तक हम सभी टेम्पलेट प्रोसेस करते हैं",
+  "Processing your file": "आपकी फ़ाइल प्रोसेस हो रही है",
+  "Bulk apply cancelled. The upload may still finish in the background.": "बल्क अप्लाई रद्द किया गया। अपलोड पृष्ठभूमि में अब भी पूरा हो सकता है।",
   "Please go through all the templates below": "कृपया नीचे दिए गए सभी टेम्पलेट देखें",
   "Sync HSM": "HSM सिंक करें",
   "Sorry, failed to sync HSM updates. Please try again.": "क्षमा करें, HSM अपडेट सिंक करने में विफल। कृपया पुनः प्रयास करें।",
